+++ .gitignore
@@ -1,2 +1,10 @@
+.DS_Store
 node_modules
-.vscode
+/build
+/.svelte-kit
+/package
+.env
+.env.*
+!.env.example
+vite.config.js.timestamp-*
+vite.config.ts.timestamp-*
