+++ src/lib/AdvancedCitation.svelte
@@ -1,1 +1,18 @@
-<span class="flex w-full items-center rounded bg-blue-100 p-6 justify-center text-3xl font-medium text-blue-800">Coming Soon</span>
+<script>
+	let citationType = "Family Tree Personal";
+	import Census from './basic_citations/Census.svelte';
+	import Church from './basic_citations/Church.svelte';
+	import CivilRecord from './basic_citations/CivilRecord.svelte';
+	import FamilyTreeHousehold from './basic_citations/FamilyTreeHousehold.svelte';
+	import FamilyTreePersonal from './basic_citations/FamilyTreePersonal.svelte'
+	import PersonalKnowledge from './basic_citations/PersonalKnowledge.svelte';
+</script>
+
+<div class="mx-auto max-w-7xl px-4 sm:px-6 md:px-8">
+	<h1 class="text-2xl font-semibold text-gray-900">Advanced Citation Builder</h1>
+</div>
+<div class="mx-auto max-w-7xl px-4 sm:px-6 md:px-8">
+	<div class="py-4">
+		<h1>Add stuff here</h1>
+	</div>
+</div>
