+++ src/routes/+page.svelte
@@ -1,0 +1,9 @@
+<script>
+	import Sidebar from '../lib/Sidebar.svelte';
+	import BasicCitation from '../lib/BasicCitation.svelte'
+</script>
+
+<Sidebar currPage="home">
+	<!-- <CitationContainer slot="container" /> -->
+	<BasicCitation slot="container"/>
+</Sidebar>
