+++ src/routes/advanced/+page.svelte
@@ -1,0 +1,8 @@
+<script>
+	import Sidebar from '../../lib/Sidebar.svelte';
+	import AdvancedCitation from '../../lib/AdvancedCitation.svelte'
+</script>
+
+<Sidebar currPage="advanced">
+	<AdvancedCitation slot="container"/>
+</Sidebar>
