+++ src/routes/soundex/+page.svelte
@@ -1,0 +1,4 @@
+<main>
+	<h1 class="text-3xl text-cyan-500">This is a test</h1>	
+</main>
+
