+++ tailwind.config.js
@@ -1,5 +1,5 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
+export default {
   content: ['./src/**/*.{html,js,svelte,ts}'],
   theme: {
     extend: {},
@@ -8,3 +8,4 @@ module.exports = {
     require('@tailwindcss/forms'),
   ],
 }
+
